+++ doc/a7r_manual.txt
@@ -1,0 +1,116 @@
+Available calls:
+getVersions
+getMethodTypes
+getApplicationInfo
+getAvailableApiList
+getEvent
+actTakePicture
+stopRecMode
+startLiveview
+stopLiveview
+startLiveviewWithSize
+actHalfPressShutter
+cancelHalfPressShutter
+setSelfTimer
+getSelfTimer
+getAvailableSelfTimer
+getSupportedSelfTimer
+getSupportedContShootingMode
+getSupportedContShootingSpeed
+getExposureMode
+getSupportedExposureMode
+setExposureCompensation
+getExposureCompensation
+getAvailableExposureCompensation
+getSupportedExposureCompensation
+setFNumber
+getFNumber
+getAvailableFNumber
+getSupportedFNumber
+setIsoSpeedRate
+getIsoSpeedRate
+getAvailableIsoSpeedRate
+getSupportedIsoSpeedRate
+getLiveviewSize
+getAvailableLiveviewSize
+getSupportedLiveviewSize
+setPostviewImageSize
+getPostviewImageSize
+getAvailablePostviewImageSize
+getSupportedPostviewImageSize
+getSupportedProgramShift
+setShootMode
+getShootMode
+getAvailableShootMode
+getSupportedShootMode
+setShutterSpeed
+getShutterSpeed
+getAvailableShutterSpeed
+getSupportedShutterSpeed
+setTouchAFPosition
+getTouchAFPosition
+setWhiteBalance
+getWhiteBalance
+getSupportedWhiteBalance
+getAvailableWhiteBalance
+getSupportedFlashMode
+setFocusMode
+getFocusMode
+getAvailableFocusMode
+getSupportedFocusMode
+setZoomSetting
+getAvailableZoomSetting
+getZoomSetting
+getSupportedZoomSetting
+getStorageInformation
+setLiveviewFrameInfo
+getLiveviewFrameInfo
+--
+getSupportedSelfTimer :
+{'result': [[0, 2, 10]], 'id': 1}
+
+getSupportedContShootingMode :
+{'result': [{'candidate': ['Single']}], 'id': 1}
+
+getSupportedContShootingSpeed :
+{'result': [{'candidate': []}], 'id': 1}
+
+getSupportedExposureMode :
+{'result': [['Program Auto', 'Aperture', 'Shutter', 'Manual', 'Intelligent Auto']], 'id': 1}
+
+getSupportedExposureCompensation :
+{'result': [[15], [-15], [1]], 'id': 1}
+
+getSupportedFNumber :
+{'result': [['1.8', '2.0', '2.2', '2.5', '2.8', '3.2', '3.5', '4.0', '4.5', '5.0', '5.6', '6.3', '7.1', '8.0', '9.0', '10', '11', '13', '14', '16', '18', '20', '22']], 'id': 1}
+
+getSupportedIsoSpeedRate :
+{'result': [['AUTO', '50', '64', '80', '100', '125', '160', '200', '250', '320', '400', '500', '640', '800', '1000', '1250', '1600', '2000', '2500', '3200', '4000', '5000', '6400', '8000', '10000', '12800', '16000', '20000', '25600']], 'id': 1}
+
+getSupportedLiveviewSize :
+{'result': [['L', 'M']], 'id': 1}
+
+getSupportedPostviewImageSize :
+{'result': [['Original', '2M']], 'id': 1}
+
+getSupportedProgramShift :
+{'result': [[]], 'id': 1}
+
+getSupportedShootMode :
+{'result': [['still']], 'id': 1}
+
+getSupportedShutterSpeed :
+{'result': [['BULB', '30"', '25"', '20"', '15"', '13"', '10"', '8"', '6"', '5"', '4"', '3.2"', '2.5"', '2"', '1.6"', '1.3"', '1"', '0.8"', '0.6"', '0.5"', '0.4"', '1\\/3', '1\\/4', '1\\/5', '1\\/6', '1\\/8', '1\\/10', '1\\/13', '1\\/15', '1\\/20', '1\\/25', '1\\/30', '1\\/40', '1\\/50', '1\\/60', '1\\/80', '1\\/100', '1\\/125', '1\\/160', '1\\/200', '1\\/250', '1\\/320', '1\\/400', '1\\/500', '1\\/640', '1\\/800', '1\\/1000', '1\\/1250', '1\\/1600', '1\\/2000', '1\\/2500', '1\\/3200', '1\\/4000', '1\\/5000', '1\\/6400', '1\\/8000']], 'id': 1}
+
+getSupportedWhiteBalance :
+{'result': [[{'colorTemperatureRange': [], 'whiteBalanceMode': 'Auto WB'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Daylight'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Shade'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Cloudy'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Incandescent'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Fluorescent: Warm White (-1)'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Fluorescent: Cool White (0)'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Fluorescent: Day White (+1)'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Fluorescent: Daylight (+2)'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Flash'}, {'colorTemperatureRange': [9900, 2500, 100], 'whiteBalanceMode': 'Color Temperature'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Custom 1'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Custom 2'}, {'colorTemperatureRange': [], 'whiteBalanceMode': 'Custom 3'}]], 'id': 1}
+
+getSupportedFlashMode :
+{'result': [['off', 'auto', 'on', 'slowSync', 'rearSync', 'wireless']], 'id': 1}
+
+getSupportedFocusMode :
+{'result': [['AF-S', 'AF-C', 'DMF', 'MF']], 'id': 1}
+
+getSupportedZoomSetting :
+{'result': [{'candidate': ['Optical Zoom Only', 'On:Clear Image Zoom', 'On:Digital Zoom']}], 'id': 1}
+
